Separate DTrace module
