+++ lib/query.js
@@ -13,25 +13,8 @@ var helpers = require('./queryhelpers');
 var Types = require('./schema/index');
 var Document = require('./document');
 var QueryStream = require('./querystream');
+var dtrace = require('./dtrace');
 
-// DTrace
-var dtrace = require('dtrace-provider');
-var dtraceProvider = dtrace.createDTraceProvider("mongoose");
-
-var queryProbe = dtraceProvider.addProbe("query-start", "char *", "int");
-var queryProbe = dtraceProvider.addProbe("query-done", "char *", "int");
-dtraceProvider.enable();
-
-var MAX_INT = Math.pow(2, 32) - 1;
-var DTRACE_ID = 0;
-
-function nextDTraceID() {
-    if (++DTRACE_ID >= MAX_INT) {
-                DTRACE_ID = 1;
-    }
-
-    return (DTRACE_ID);
-};
 
 /**
  * Query constructor used for building queries.
@@ -1096,8 +1079,8 @@ function completeMany (model, docs, fields, self, pop, promise) {
  */
 
 Query.prototype.findOne = function (conditions, fields, options, callback) {
-  var queryId = nextDTraceID();
-  dtraceProvider.fire("query-start", function(p) { 
+  var queryId = dtrace.nextDTraceID();
+  dtrace.dtraceProvider.fire("query-start", function(p) { 
     return ["findOne", queryId];
   });
 
@@ -1170,7 +1153,7 @@ Query.prototype.findOne = function (conditions, fields, options, callback) {
     if (!doc) return promise.complete(null);
 
     if (!options.populate) {
-      dtraceProvider.fire("query-done", function(p) { 
+      dtrace.dtraceProvider.fire("query-done", function(p) { 
         return ["findOne", queryId];
       });
       return true === options.lean
@@ -1182,7 +1165,7 @@ Query.prototype.findOne = function (conditions, fields, options, callback) {
     self.model.populate(doc, pop, function (err, doc) {
       if (err) return promise.error(err);
 
-      dtraceProvider.fire("query-done", function(p) { 
+      dtrace.dtraceProvider.fire("query-done", function(p) { 
         return ["findOne", queryId];
       });
 
@@ -1516,8 +1499,8 @@ Query.prototype._findAndModify = function (type, callback) {
     throw new Error("Expected callback in _findAndModify");
   }
 
-  var queryId = nextDTraceID();
-  dtraceProvider.fire("query-start", function(p) { 
+  var queryId = dtrace.nextDTraceID();
+  dtrace.dtraceProvider.fire("query-start", function(p) { 
     return ["findAndModify", queryId];
   });
 
@@ -1584,7 +1567,7 @@ Query.prototype._findAndModify = function (type, callback) {
     }
 
     if (!options.populate) {
-      dtraceProvider.fire("query-done", function(p) { 
+      dtrace.dtraceProvider.fire("query-done", function(p) { 
         return ["findAndModify", queryId];
       });
       return true === options.lean
@@ -1596,7 +1579,7 @@ Query.prototype._findAndModify = function (type, callback) {
     self.model.populate(doc, pop, function (err, doc) {
       if (err) return promise.error(err);
 
-      dtraceProvider.fire("query-done", function(p) { 
+      dtrace.dtraceProvider.fire("query-done", function(p) { 
         return ["findAndModify", queryId];
       });
       return true === options.lean
